+++ _example/object/getAnonymous.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"net/url"
+	"os"
 	"strings"
 
 	"io/ioutil"
@@ -27,7 +28,7 @@ func upload(c *cos.Client, name string) {
 }
 
 func main() {
-	u, _ := url.Parse("https://test-1253846586.cn-north.myqcloud.com")
+	u, _ := url.Parse(os.Getenv("COS_BUCKET_URL"))
 	b := &cos.BaseURL{BucketURL: u}
 	c := cos.NewClient(b, nil)