+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.8.0] (2018-05-26)
+
+### 新增
+
+* 新增 `func NewBaseURL(bucketURL string) (u *BaseURL, err error)` (via [91f7759])
+
+### 变更
+
+* `NewBucketURL` 函数使用新的 URL 域名规则。(via [7dcd701])     
+  影响：如果有使用 `NewBucketURL` 函数生成 bucketURL 的话，使用时需要使用新的 Region 名称，
+  详见 https://cloud.tencent.com/document/product/436/6224 ，未使用 `NewBucketURL` 函数不受影响
+
+
 ## [0.7.0] (2017-12-23)
 
 ### 新增
@@ -49,7 +62,7 @@
 
 ## [0.2.0] (2017-06-10)
 
-### 修改
+### 变更
 
 * 调用 bucket 相关 API 时不再需要 bucket 参数, 把参数移到 service 中
 * 把参数 signStartTime, signEndTime, keyStartTime, keyEndTime 合并为 authTime
@@ -63,9 +76,13 @@
 * 完成大部分 Bucket API(还剩一个 Put Bucket Lifecycle)
 
 
+[0.8.0]: https://github.com/mozillazg/go-cos/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/mozillazg/go-cos/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/mozillazg/go-cos/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/mozillazg/go-cos/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/mozillazg/go-cos/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/mozillazg/go-cos/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/mozillazg/go-cos/compare/v0.1.0...v0.2.0
+
+[91f7759]: https://github.com/mozillazg/go-cos/commit/91f7759958f9631e8997f47d30ae4044455fc971
+[7dcd701]: https://github.com/mozillazg/go-cos/commit/7dcd701975f483d57525b292ab31d0f9a6c8866c